requirements.txt
dotenv
requests
python-dotenv
markdown
espn_api
google-genai
nba_api